Split out spans into a new module
The span iterators and tests took up a significant portion of the
syntax module but were fairly self-contained.

In addition, the Span type was an informal

    (usize, std::ops::Range<usize>)

which was a bit tricky to deal with:

* the type was long to write out for all producers/consumers
* std::ops::Range<usize> does not implement Copy

Plus we don't end up using any Range methods except its Ord/PartialOrd
implementations.

Given its first-class usage for diagnostics and selections, it seems
appropriate to separate it into its own struct type and module.
